+++ src/bin/s3sync/cli/mod.rs
@@ -1,5 +1,4 @@
 use anyhow::{Result, anyhow};
-use cfg_if::cfg_if;
 use s3sync::Config;
 use s3sync::callback::user_defined_event_callback::UserDefinedEventCallback;
 use s3sync::callback::user_defined_filter_callback::UserDefinedFilterCallback;
@@ -32,7 +31,6 @@ pub async fn run(mut config: Config) -> Result<()> {
         let cancellation_token = create_pipeline_cancellation_token();
 
         // Note: Each type of callback is registered only once.
-
         // The user-defined event callback is disabled by default.
         let user_defined_event_callback = UserDefinedEventCallback::new();
         if user_defined_event_callback.is_enabled() {
@@ -41,30 +39,6 @@ pub async fn run(mut config: Config) -> Result<()> {
             config
                 .event_manager
                 .register_callback(EventType::ALL_EVENTS, user_defined_event_callback);
-        } else {
-            cfg_if! {
-                if #[cfg(feature = "lua_support")] {
-                    #[allow(clippy::collapsible_else_if)]
-                    if let Some(event_callback_lua_script) = config.event_callback_lua_script.as_ref() {
-                        let mut lua_event_callback =
-                            s3sync::callback::lua_event_callback::LuaEventCallback::new(
-                                config.lua_vm_memory_limit,
-                                config.allow_lua_os_library,
-                                config.allow_lua_unsafe_vm
-                            );
-                        if let Err(e) = lua_event_callback
-                            .load_and_compile(event_callback_lua_script.as_str())
-                            .await
-                        {
-                            error!("Failed to load and compile Lua script event callback: {}",e);
-                            return Err(anyhow!("Failed to load and compile Lua script event callback"));
-                        }
-
-                        // Lua event callback is registered for all events.
-                        config.event_manager.register_callback(EventType::ALL_EVENTS, lua_event_callback);
-                    }
-                }
-            }
         }
 
         // The user-defined filter callback is disabled by default.
@@ -76,28 +50,6 @@ pub async fn run(mut config: Config) -> Result<()> {
                 .filter_config
                 .filter_manager
                 .register_callback(user_defined_filter_callback);
-        } else {
-            cfg_if! {
-                if #[cfg(feature = "lua_support")] {
-                    #[allow(clippy::collapsible_else_if)]
-                    if let Some(filter_callback_lua_script) = config.filter_callback_lua_script.as_ref() {
-                        let mut lua_filter_callback =
-                            s3sync::callback::lua_filter_callback::LuaFilterCallback::new(
-                                config.lua_vm_memory_limit,
-                                config.allow_lua_os_library,
-                                config.allow_lua_unsafe_vm
-                            );
-                        if let Err(e) = lua_filter_callback
-                            .load_and_compile(filter_callback_lua_script.as_str())
-                            .await
-                        {
-                            error!("Failed to load and compile Lua script filter callback: {}",e);
-                            return Err(anyhow!("Failed to load and compile Lua script filer callback"));
-                        }
-                        config.filter_config.filter_manager.register_callback(lua_filter_callback);
-                    }
-                }
-            }
         }
 
         // The user-defined preprocess callback is disabled by default.
@@ -108,28 +60,6 @@ pub async fn run(mut config: Config) -> Result<()> {
             config
                 .preprocess_manager
                 .register_callback(user_defined_preprocess_callback);
-        } else {
-            cfg_if! {
-                if #[cfg(feature = "lua_support")] {
-                    #[allow(clippy::collapsible_else_if)]
-                    if let Some(preprocess_callback_lua_script) = config.preprocess_callback_lua_script.as_ref() {
-                        let mut lua_preprocess_callback =
-                            s3sync::callback::lua_preprocess_callback::LuaPreprocessCallback::new(
-                                config.lua_vm_memory_limit,
-                                config.allow_lua_os_library,
-                                config.allow_lua_unsafe_vm
-                            );
-                        if let Err(e) = lua_preprocess_callback
-                            .load_and_compile(preprocess_callback_lua_script.as_str())
-                            .await
-                        {
-                            error!("Failed to load and compile Lua script preprocess callback: {}",e);
-                            return Err(anyhow!("Failed to load and compile Lua script preprocess callback"));
-                        }
-                        config.preprocess_manager.register_callback(lua_preprocess_callback);
-                    }
-                }
-            }
         }
 
         ctrl_c_handler::spawn_ctrl_c_handler(cancellation_token.clone());
@@ -212,6 +142,7 @@ mod tests {
     }
 
     #[tokio::test]
+    #[cfg(feature = "lua_support")]
     async fn event_callback_lua_script_error() {
         init_dummy_tracing_subscriber();
 
@@ -223,13 +154,12 @@ mod tests {
             "--event-callback-lua-script",
             "./test_data/script/invalid_script.lua",
         ];
-        let config = Config::try_from(parse_from_args(args).unwrap()).unwrap();
-
-        let result = run(config).await;
-        assert!(result.is_err());
+        let config = Config::try_from(parse_from_args(args).unwrap());
+        assert!(config.is_err());
     }
 
     #[tokio::test]
+    #[cfg(feature = "lua_support")]
     async fn filter_callback_lua_script_error() {
         init_dummy_tracing_subscriber();
 
@@ -241,13 +171,12 @@ mod tests {
             "--filter-callback-lua-script",
             "./test_data/script/invalid_script.lua",
         ];
-        let config = Config::try_from(parse_from_args(args).unwrap()).unwrap();
-
-        let result = run(config).await;
-        assert!(result.is_err());
+        let config = Config::try_from(parse_from_args(args).unwrap());
+        assert!(config.is_err());
     }
 
     #[tokio::test]
+    #[cfg(feature = "lua_support")]
     async fn preprocess_callback_lua_script_error() {
         init_dummy_tracing_subscriber();
 
@@ -259,10 +188,9 @@ mod tests {
             "--preprocess-callback-lua-script",
             "./test_data/script/invalid_script.lua",
         ];
-        let config = Config::try_from(parse_from_args(args).unwrap()).unwrap();
+        let config = Config::try_from(parse_from_args(args).unwrap());
 
-        let result = run(config).await;
-        assert!(result.is_err());
+        assert!(config.is_err());
     }
 
     #[tokio::test]