+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.34.0] - 2025-08-13
+
+### Changed
+
+- [Library] Lua scripting support  
+  You can specify Lua scripting support CLI arguments in the library.
+- aws-sdk-s3 = "1.102.0"
+- Updated documentation.
+- Updated dependencies.
+
+
 ## [1.33.0] - 2025-08-12
 
 ### Changed